+++ aoe4/aoe4_build_order.py
@@ -1,5 +1,5 @@
 from aoe4.aoe4_civ_icon import aoe4_civilization_icon
-from common.build_order_tools import is_valid_resource
+from common.build_order_tools import is_valid_resource, build_order_time_to_str
 
 
 def check_valid_aoe4_build_order(data: dict, bo_name_msg: bool = False) -> (bool, str):
@@ -182,3 +182,23 @@ def get_aoe4_build_order_template() -> dict:
         'source': 'Source',
         'build_order': [get_aoe4_build_order_step()]
     }
+
+
+def evaluate_aoe4_build_order_timing(data: dict):
+    """Evaluate the time indications for an AoE4 build order.
+
+    Parameters
+    ----------
+    data    Data of the build order (will be updated).
+    """
+    # creation times [sec]
+    villager_time: int = 20
+
+    if 'build_order' not in data:
+        print('The \'build_order\' field is missing from data when evaluating the timing.')
+        return
+
+    build_order_data = data['build_order']
+    for step in build_order_data:
+        villager_count = step['villager_count']
+        step['time'] = build_order_time_to_str(villager_time * villager_count)